chg: revert moving add_choices_info_to_question out of xls2json.py
- moved out mainly to support importing from other modules that xls2json
  depends on, but it turned out that it was not needed. In particular,
  as noted in the function body, it supports various legacy scenarios
  which are not relevant to newer features such as itemset referencing
  in range or geo parameters.
