+++ pyxform/validators/pyxform/choices.py
@@ -1,8 +1,5 @@
-from typing import Any
-
 from pyxform import constants as co
 from pyxform.errors import ErrorCode, PyXFormError
-from pyxform.validators.pyxform.pyxform_reference import has_pyxform_reference
 
 
 def validate_headers(
@@ -70,53 +67,3 @@ def validate_and_clean_choices(
                 option.pop(invalid_header, None)
             option.pop("__row", None)
     return choices
-
-
-def add_choices_info_to_question(
-    question: dict[str, Any],
-    list_name: str,
-    choices: dict[str, list],
-    choice_filter: str | None = None,
-    file_extension: str | None = None,
-):
-    """
-    Add choices-related info to the question dict, e.g. itemset, list_name, choices, etc.
-
-    :param question: A dict with question details.
-    :param list_name: The choice list name for the question.
-    :param choices: The available choices in the survey.
-    :param choice_filter: The question's choice_filter, if any.
-    :param file_extension: The question's external select file_extension, if any.
-    :return: The updated question dict.
-    """
-    if choice_filter is None:
-        choice_filter = ""
-    if file_extension is None:
-        file_extension = ""
-
-    question[co.ITEMSET] = list_name
-
-    if choice_filter:
-        # External selects e.g. type = "select_one_external city".
-        if question[co.TYPE] == co.SELECT_ONE_EXTERNAL:
-            question["query"] = list_name
-        elif choices.get(list_name):
-            # Reference to list name for data dictionary tools (ilri/odktools).
-            question[co.LIST_NAME_U] = list_name
-            # Copy choices for data export tools (onaio/onadata).
-            # TODO: could onadata use the list_name to look up the list for
-            #  export, instead of pyxform internally duplicating choices data?
-            question[co.CHOICES] = choices[list_name]
-    elif not (
-        # Select with randomized choices.
-        (
-            co.ParametersSelect.RANDOMIZE in question[co.PARAMETERS]
-            and question[co.PARAMETERS][co.ParametersSelect.RANDOMIZE] == "true"
-        )
-        # Select from file e.g. type = "select_one_from_file cities.xml".
-        or file_extension in co.EXTERNAL_INSTANCE_EXTENSIONS
-        # Select from previous answers e.g. type = "select_one ${q1}".
-        or has_pyxform_reference(list_name)
-    ):
-        question[co.LIST_NAME_U] = list_name
-        question[co.CHOICES] = choices[list_name]
